Refactor controller into many sub controllers
MOAR WIP

all tests passing :fireworks:

downcase utils

downcase utils in another place

include logic about redirects
