+++ index.js
@@ -3,7 +3,7 @@ var pkg = require('./package.json')
 var provider = {
   plugin_name: 'agol',
   hosts: true,
-  controller: require('./controller'),
+  controller: require('./controllers'),
   routes: require('./routes'),
   model: require('./models/agol.js'),
   status: {
