+++ routes/index.js
@@ -1,28 +1,24 @@
 module.exports = {
-  'post /agol': 'register',
-  'get /agol': 'list',
-  'get /agol/queue/length': 'getQueueLength',
-  'get /agol/queue/workers': 'getQueueWorkers',
-  'get /agol/queue/working': 'getQueueWorkingCount',
-  'get /agol/queue/clearFailed': 'clearFailedJobs',
-  'get /agol/:id': 'find',
+  'post /agol': 'hosts',
+  'get /agol': 'hosts',
+  'get /agol/:id': 'hosts',
+  'delete /agol/:id': 'hosts',
+  'get /agol/queue/:action': 'queue',
   'get /agol/:id/:item/*': 'setHost',
   'put /agol/:id/:item/*': 'setHost',
   'post /agol/:id/:item/*': 'setHost',
   'post /agol/:id/bulk': 'setHost',
-  'post /agol/:id/bulk/import': 'bulkImport',
-  'post /agol/:id/bulk/export': 'bulkExport',
+  'post /agol/:id/bulk/:action': 'bulk',
   'get /agol/:id/:item': 'getInfo',
   // this method needs to come before layer/format
   'get /agol/:id/:item/FeatureServer': 'featureserver',
   // end method that needs to be before layer/format
-  'delete /agol/:id': 'del',
-  'delete /agol/:id/:item/:layer': 'deleteItemData',
   'get /agol/:id/:item/:layer/geohash': 'getGeohash',
   'get /agol/:id/:item.:format': 'getResource',
   'get /agol/:id/:item/:layer.:format': 'getResource',
   'get /agol/:id/:item/:layer': 'getInfo',
-  'get /agol/:id/:item/:layer/drop': 'dropResource',
+  'get /agol/:id/:item/:layer/drop': 'drop',
+  'delete /agol/:id/:item/:layer': 'drop',
   'get /agol/:id/:item/:layer/tiles/:z/:x/:y.:format': 'tiles',
   'get /agol/:id/:item/:layer/tiles/:z/:x/:y': 'tiles',
   'get /agol/:id/:item/FeatureServer/geohash': 'getGeohash',
@@ -30,8 +26,7 @@ module.exports = {
   'get /agol/:id/:item/FeatureServer/:layer/geohash': 'getGeohash',
   'get /agol/:id/:item/FeatureServer/:layer/:method': 'featureserver',
   'post /agol/:id/:item/FeatureServer/:layer/:method': 'featureserver',
-  'get /agol/:id/:item/:layer/expiration': 'getExpiration',
-  'put /agol/:id/:item/:layer/expiration': 'setExpiration',
-  'post /agol/:id/:item/:layer/expiration': 'setExpiration',
-  'get /test': 'testRoute'
+  'get /agol/:id/:item/:layer/expiration': 'expiration',
+  'put /agol/:id/:item/:layer/expiration': 'expiration',
+  'post /agol/:id/:item/:layer/expiration': 'expiration'
 }
