+++ test/model-test.js
@@ -24,99 +24,6 @@ koop.fs = new LocalFs({rootDir: path.join(__dirname, 'data')})
 var agol = require('../models/agol.js')(koop)
 
 describe('AGOL Model', function () {
-  describe('getting info', function () {
-    afterEach(function (done) {
-      agol.cache.getInfo.restore()
-      done()
-    })
-    it('should return status: unavailable when the resource is not cached', function (done) {
-      sinon.stub(agol.cache, 'getInfo', function (key, callback) {
-        callback(new Error('Resource not found'))
-      })
-
-      agol.getInfo({key: 'test'}, function (err, info) {
-        should.not.exist(err)
-        info.status.should.equal('Unavailable')
-        done()
-      })
-    })
-
-    it('should return with status: Cached when the resource is not expired', function (done) {
-      sinon.stub(agol.cache, 'getInfo', function (key, callback) {
-        callback(null, {status: 'Cached', type: 'Feature Service', version: 3})
-      })
-
-      sinon.stub(agol.cache, 'checkExpiration', function (info, layer, callback) {
-        callback(null, false, {status: 'Cached', type: 'Feature Service', version: 3})
-      })
-
-      agol.getInfo({key: 'test'}, function (err, info) {
-        should.not.exist(err)
-        info.status.should.equal('Cached')
-        agol.cache.checkExpiration.restore()
-        done()
-      })
-    })
-
-    it('should return with status: Processing when a feature service resource is processing', function (done) {
-      sinon.stub(agol.cache, 'getInfo', function (key, callback) {
-        callback(null, {status: 'Processing', version: 3, type: 'Feature Service'})
-      })
-
-      agol.getInfo({key: 'test'}, function (err, info) {
-        should.not.exist(err)
-        info.status.should.equal('Processing')
-        done()
-      })
-    })
-
-    it('should return with status: Expired with a csv item has been modified', function (done) {
-      sinon.stub(agol.cache, 'getInfo', function (key, callback) {
-        callback(null, {status: 'Cached', retrieved_at: Date.now(), version: 3, type: 'CSV', url: 'foo'})
-      })
-
-      sinon.stub(agol.portal, 'getItem', function (host, item, callback) {
-        callback(null, {modified: Date.now() + 1000})
-      })
-
-      agol.getInfo({key: 'test'}, function (err, info) {
-        should.not.exist(err)
-        info.status.should.equal('Expired')
-        agol.portal.getItem.restore()
-        done()
-      })
-    })
-
-    it('should call updateResource when resource is expired', function (done) {
-      sinon.stub(agol.cache, 'getInfo', function (key, callback) {
-        callback(null, {status: 'Expired', type: 'Feature Service', version: 3})
-      })
-
-      sinon.stub(agol.cache, 'checkExpiration', function (info, layer, callback) {
-        callback(null, {status: 'Expired', type: 'Feature Service', version: 3})
-      })
-
-      sinon.stub(agol, 'updateResource', function (info, options, callback) {
-        callback(null)
-        execTest()
-      })
-
-      agol.getInfo({key: 'test'}, function (err, info) {
-        should.not.exist(err)
-        info.status.should.equal('Expired')
-      })
-
-      function execTest () {
-        agol.cache.checkExpiration.called.should.equal(true)
-        agol.updateResource.called.should.equal(true)
-
-        agol.cache.checkExpiration.restore()
-        agol.updateResource.restore()
-        done()
-      }
-    })
-  })
-
   describe('remove items', function () {
     before(function (done) {
       sinon.stub(koop.cache, 'remove', function (host, itemid, opts, callback) {
